Split liquidity source context mapping
